move estimateTimeFormat to common
