+++ projects/common/src/lib/util/conversions.ts
@@ -5,6 +5,7 @@ import OlGeometry from 'ol/geom/Geometry';
 import {Extent as OlExtent} from 'ol/extent';
 import {Observable, ReplaySubject} from 'rxjs';
 import {BoundingBox2D as BBoxDict, ResponseError} from '@geoengine/openapi-client';
+import {Time} from '../time/time.model';
 
 /**
  * Converts an `OlExtent` to an extent as a tuple of four numbers.
@@ -101,4 +102,36 @@ export async function errorToText(error: any, defaultMessage: string): Promise<s
     const errorMessage = errorJson.message ?? defaultMessage;
 
     return errorMessage;
+}
+
+// we use this non-breaking hyphen to avoid line breaks in the time format
+export const NON_BREAKING_HYPHEN: string = '‑';
+
+/**
+ * Checks if the `timeSteps` contain information about months or days.
+ * For instance, for the steps 2019-01-01, 2019-02-01, 2019-03-01, the format would be 'YYYY-MM'.
+ */
+export function estimateTimeFormat(timeSteps: Array<Time>): string {
+    let useMonth = false;
+    let useDay = false;
+    for (const timeStep of timeSteps) {
+        // in range [1, 31]
+        if (timeStep.start.date() > 1) {
+            useDay = useMonth = true;
+            break;
+        }
+        // in range [0, 11]
+        if (timeStep.start.month() > 0) {
+            useMonth = true;
+            // we cannot break, it could still be that we need to use days
+        }
+    }
+
+    if (useDay) {
+        return `YYYY${NON_BREAKING_HYPHEN}MM${NON_BREAKING_HYPHEN}DD`;
+    } else if (useMonth) {
+        return `YYYY${NON_BREAKING_HYPHEN}MM`;
+    } else {
+        return 'YYYY'; // default: yearly
+    }
 }
