+++ projects/dashboards/ebv-analyzer/src/app/data-selection.service.ts
@@ -3,7 +3,7 @@ import {LoadingState, ProjectService} from '@geoengine/core';
 import {first, map, mergeMap, tap} from 'rxjs/operators';
 import {BehaviorSubject, combineLatest, Observable, of} from 'rxjs';
 import moment from 'moment';
-import {Layer, RasterLayer, Time, VectorLayer} from '@geoengine/common';
+import {estimateTimeFormat, Layer, RasterLayer, Time, VectorLayer} from '@geoengine/common';
 
 export interface DataRange {
     min: number;
@@ -120,37 +120,5 @@ export class DataSelectionService {
             mergeMap(() => this.projectService.addLayer(layer)),
             tap(() => this._polygonLayer.next(layer)),
         );
-    }
-}
-
-// we use this non-breaking hyphen to avoid line breaks in the time format
-export const NON_BREAKING_HYPHEN: string = '‑';
-
-/**
- * Checks if the `timeSteps` contain information about months or days.
- * For instance, for the steps 2019-01-01, 2019-02-01, 2019-03-01, the format would be 'YYYY-MM'.
- */
-export function estimateTimeFormat(timeSteps: Array<Time>): string {
-    let useMonth = false;
-    let useDay = false;
-    for (const timeStep of timeSteps) {
-        // in range [1, 31]
-        if (timeStep.start.date() > 1) {
-            useDay = useMonth = true;
-            break;
-        }
-        // in range [0, 11]
-        if (timeStep.start.month() > 0) {
-            useMonth = true;
-            // we cannot break, it could still be that we need to use days
-        }
-    }
-
-    if (useDay) {
-        return `YYYY${NON_BREAKING_HYPHEN}MM${NON_BREAKING_HYPHEN}DD`;
-    } else if (useMonth) {
-        return `YYYY${NON_BREAKING_HYPHEN}MM`;
-    } else {
-        return 'YYYY'; // default: yearly
     }
 }
